patch 0.2.0: Update hardware files to KiCad 5.
This patch brought a firmware part of the project to perform any tests
only. It draws two different pictures on an OLED display and counts
frames per second.

Schematic was updated to include a new DC-DC buck converter (AP3401) to
produce 3.3V from both USB and a battery. Choosing this converter is a
trade-off between an expensive buck-boost converter which can operate
correctly at 3.2V voltage level of almost depleted battery and a cheap buck
converter which produces 3.3V from the input voltage which should be above
the required level.

There is also a python plugin in hardware/plugins directory to generate a BOM
html file. It can be added to and called from KiCad directly.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>