patch 0.2.0001: Adjust some footprints.
Several footprints were updated to use rounded rectangles instead of
rectangles with sharp corners as pads.

BOM generator script updated to include a revision into a BOM html file.
It is useful to understand which components are required for a specific
version of the device.

There is a new TSOT26 footprint created and added especially for AP3401
DC-DC converter.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>