patch 3.1: More checkpoints and 4-wire SPI.
This is the first commit after version adjustment and in a new
current/3.0 branch. All of the changes in the third iteration of the
Xling PCB will find their place in this branch.

Starting from Xling PCB v3.1, new checkpoints are available:

    +3.3V - output voltage of the power supply block and AP3401
            particularly;
    VBAT - Li-Ion battery voltage;
    SCK - SCK pin of the SPI;
    MOSI - MOSI pin of the SPI;
    CS - Chip-select pin of the OLED display;
    DC - Data/Command pin of the OLED display;
    RST - Reset pin of the OLED display;
    GND - Ground of the PCB.

And the display is connected via 4-wire SPI from now on.

Signed-off-by: Dmitry Salychev <dsl@mcusim.org>