Road to OSHWA certification.
This is the first commit in the current/4.0 branch which brings a lot of
changes which are supposed to gather all files in one place and open a road
to OSHWA certification.

Structure of the repository changed. It contains all of the necessary
design documents, firmware source code and other files to let everyone
make the device.

Repository includes the firmware of the version 0.4.12 which
means that two critical bugs when Xling hangs fixed.

KiCad project, schematic and board designs are distrubuted under the
terms of CERN Open Hardware Licence Version 2 - Weakly Reciprocal
starting from hardware revision 4.